+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Cloudkill checked save at wrong point in the turn.
 * Fixcritters function removed as no longer needed.
 * Experimental bulk import mode has been removed due to duplication of data issues.
+* Legendary resources no longer tracked by default on bar2 monsters with legendary actions.
 
 # 3.2.10