+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Prevent spell duplication for some feat choices.
 * Effects: Tweaks to DamageOnlyWorkflow macros to support levels auto cover and midi-qol when cover is enabled. @Chris#8375
 * Character Updates: if the death saves were cleared rather than set to 0, update to DDB would fail.
+* Extras Import: Would duplicate folder, extra or move extras to root directory.
 
 # 3.2.12